+++ src/shmif/tui/lua/nbio.c
@@ -37,11 +37,9 @@
 	#define lua_rawlen(x, y) lua_objlen(x, y)
 #endif
 
-static void check_canary(struct nonblock_io* ib)
-{
-	if (ib->canary_pre != 0xfeedface || ib->canary_post != 0xfacefeed)
-		abort();
-}
+#ifdef WANT_ARCAN_BASE
+#define arcan_fatal(...) do { alt_fatal( __VA_ARGS__); } while(0)
+#endif
 
 static struct nonblock_io open_fds[LUACTX_OPEN_FILES];
 /* open_nonblock and similar functions need to register their fds here as they
@@ -236,7 +234,6 @@ int alt_nbio_socket(const char* path, int ns, char** out)
 int alt_nbio_process_write(lua_State* L, struct nonblock_io* ib)
 {
 	struct io_job* job = ib->out_queue;
-	check_canary(ib);
 
 	while (job){
 		ssize_t nw = write(ib->fd, &job->buf[job->ofs], job->sz - job->ofs);
@@ -481,12 +478,11 @@ static int nbio_socketaccept(lua_State* L)
 		sizeof(struct nonblock_io), ARCAN_MEM_BINDING, 0, ARCAN_MEMALIGN_NATURAL);
 
 	(*conn) = (struct nonblock_io){
-		.canary_pre = 0xfeedface,
 		.fd = newfd,
 		.mode = O_RDWR,
 		.data_handler = LUA_NOREF,
 		.write_handler = LUA_NOREF,
-		.canary_post = 0xfacefeed
+		.lfch = '\n'
 	};
 
 	if (!conn){
@@ -662,7 +658,8 @@ static int nbio_write(lua_State* L)
 }
 
 static char* nextline(struct nonblock_io* ib,
-	size_t start, bool eof, size_t* nb, size_t* step, bool* gotline)
+	size_t start, bool eof, size_t* nb, size_t* step,
+	bool* gotline, char linech)
 {
 	if (!ib->ofs)
 		return NULL;
@@ -670,7 +667,7 @@ static char* nextline(struct nonblock_io* ib,
 	*step = 0;
 
 	for (size_t i = start; i < ib->ofs; i++){
-		if (ib->buf[i] == '\n'){
+		if (ib->buf[i] == linech){
 			*nb = ib->lfstrip ? (i - start) : (i - start) + 1;
 			*step = (i - start) + 1;
 			*gotline = true;
@@ -678,22 +675,18 @@ static char* nextline(struct nonblock_io* ib,
 		}
 	}
 
-/* there might be data left or we have hit the buffering limit
- * and need to forward without waiting for a newline */
 	if (eof || (!start && ib->ofs == COUNT_OF(ib->buf))){
 		*gotline = false;
-
-		if (ib->ofs <= start){
+		if (ib->ofs < start){
 			*nb = 0;
 			*step = 0;
 			ib->ofs = 0;
-			return NULL;
 		}
 		else {
 			*nb = ib->ofs - start;
 			*step = ib->ofs - start;
 		}
-		return &ib->buf[start];
+		return ib->buf;
 	}
 
 	return NULL;
@@ -722,7 +715,6 @@ int alt_nbio_process_read(
  */
 	bool eof = false;
 	ssize_t nr = read(ib->fd, &ib->buf[ib->ofs], buf_sz - ib->ofs);
-	check_canary(ib);
 
 	if (0 == nr){
 		eof = true;
@@ -765,8 +757,6 @@ int alt_nbio_process_read(
  * 3. forward to the callback at -1.
  */
 #define SLIDE(X) do{\
-	if (ib->ofs < ci)\
-		ib->ofs = ci;\
 	memmove(ib->buf, &ib->buf[ci], ib->ofs - ci);\
 	ib->ofs -= ci;\
 }while(0)
@@ -783,7 +773,7 @@ int alt_nbio_process_read(
 		bool cancel = false;
 		while (
 			!cancel &&
-			(ch = nextline(ib, ci, eof, &len, &step, &gotline))){
+			(ch = nextline(ib, ci, eof, &len, &step, &gotline, ib->lfch))){
 			lua_pushvalue(L, -1);
 			lua_pushlstring(L, ch, len);
 			lua_pushboolean(L, eof && !gotline);
@@ -794,7 +784,6 @@ int alt_nbio_process_read(
 		}
 
 		SLIDE();
-		check_canary(ib);
 
 		lua_pushnil(L);
 		lua_pushboolean(L, !eof);
@@ -814,7 +803,7 @@ int alt_nbio_process_read(
 
 	while (
 			count &&
-			(ch = nextline(ib, ci, eof, &len, &step, &gotline))){
+			(ch = nextline(ib, ci, eof, &len, &step, &gotline, ib->lfch))){
 			if (eof && len == 0 && step == 0)
 				break;
 
@@ -824,16 +813,13 @@ int alt_nbio_process_read(
 			count--;
 			ci += step;
 		}
-
 		SLIDE();
-		check_canary(ib);
-
 		lua_pushnil(L);
 		lua_pushboolean(L, !eof);
 		return 2;
 	}
 	else {
-		if ((ch = nextline(ib, 0, eof, &len, &step, &gotline))){
+		if ((ch = nextline(ib, 0, eof, &len, &step, &gotline, ib->lfch))){
 			lua_pushlstring(L, ch, len);
 			memmove(ib->buf, &ib->buf[step], buf_sz - step);
 			ib->ofs -= step;
@@ -841,7 +827,6 @@ int alt_nbio_process_read(
 		else
 			lua_pushnil(L);
 
-		check_canary(ib);
 		lua_pushboolean(L, !eof);
 		return 2;
 	}
@@ -854,6 +839,11 @@ static int nbio_lf(lua_State* L)
 	struct nonblock_io* ir = *ib;
 
 	ir->lfstrip = luaL_optbnumber(L, 2, 0);
+
+	if (lua_type(L, 3) == LUA_TSTRING){
+		const char* ch = lua_tostring(L, 3);
+		ir->lfch = ch[0];
+	}
 
 	LUA_ETRACE("open_nonblock:lf_strip", NULL, 0);
 }
@@ -886,26 +876,42 @@ static int opennonblock_tgt(lua_State* L, bool wr)
 	if (vobj->feed.state.tag != ARCAN_TAG_FRAMESERV)
 		arcan_fatal("open_nonblock(tgt), target must be a valid frameserver.");
 
-	int outp[2];
-	if (-1 == pipe(outp)){
-		arcan_warning("open_nonblock(tgt), pipe-pair creation failed: %d\n", errno);
-		return 0;
-	}
-
+/* overloaded form:
+ *  open_nonblock(vid, r | w, type, nbio_ud)
+ *
+ *  This takes an existing userdata, extracts the descriptor and sends to the
+ *  target, while disassociating the descriptor from the argument source.
+ */
 	const char* type = luaL_optstring(L, 3, "stream");
-
-/* WRITE mode = 'INPUT' in the client space */
-	int dst = wr ? outp[0] : outp[1];
-	int src = wr ? outp[1] : outp[0];
-
-/* in any scenario where this would fail, "blocking" behavior is acceptable */
-	alt_nbio_nonblock_cloexec(src, true);
 	struct arcan_event ev = {
 		.category = EVENT_TARGET,
 		.tgt.kind = wr ? TARGET_COMMAND_BCHUNK_IN : TARGET_COMMAND_BCHUNK_OUT
 	};
 	snprintf(ev.tgt.message, COUNT_OF(ev.tgt.message), "%s", type);
+	if (lua_type(L, 4) == LUA_TUSERDATA){
+		struct nonblock_io** ibb = luaL_checkudata(L, 4, "nonblockIO");
+		struct nonblock_io* ib = *ibb;
 
+		if (ib->fd > 0){
+			platform_fsrv_pushfd(fsrv, &ev, ib->fd);
+			close(ib->fd);
+			ib->fd = -1;
+		}
+
+		return 0;
+	}
+
+/* WRITE mode = 'INPUT' in the client space */
+	int outp[2];
+	if (-1 == pipe(outp)){
+		arcan_warning("open_nonblock(tgt), pipe-pair creation failed: %d\n", errno);
+		return 0;
+	}
+	int dst = wr ? outp[0] : outp[1];
+	int src = wr ? outp[1] : outp[0];
+
+/* in any scenario where this would fail, "blocking" behavior is acceptable */
+	alt_nbio_nonblock_cloexec(src, true);
 	if (ARCAN_OK != platform_fsrv_pushfd(fsrv, &ev, dst)){
 		close(dst);
 		close(src);
@@ -921,8 +927,6 @@ static int opennonblock_tgt(lua_State* L, bool wr)
 		return 0;
 	}
 
-	conn->canary_pre = 0xfeedface;
-	conn->canary_post = 0xfacefeed;
 	conn->mode = wr ? O_WRONLY : O_RDONLY;
 	conn->fd = src;
 	conn->pending = NULL;
@@ -959,19 +963,215 @@ void alt_nbio_release()
 	}
 }
 
+struct pathfd {
+	char* path;
+	char* unlink;
+	const char* err;
+	const char* metatable;
+	int fd;
+	int wrmode;
+};
+
+static struct pathfd build_fifo_ipc(char* path, bool userns, bool expect_write)
+{
+	struct pathfd res = {
+		.path = NULL,
+		.fd = -1,
+		.err = NULL
+	};
+	int ns = userns ? RESOURCE_NS_USER : RESOURCE_APPL_TEMP;
+
+	int fl = O_NONBLOCK | O_WRONLY | O_CLOEXEC;
+	char* workpath = arcan_expand_resource(path, ns);
+	if (!workpath){
+		res.err = "Couldn't expand FIFO path";
+		return res;
+	}
+
+/* if it doesn't exist and we are the write end, create and try again */
+	struct stat fi;
+	if (-1 == stat(path, &fi)){
+		if (expect_write){
+			if (-1 == mkfifo(workpath, S_IRWXU)){
+				arcan_mem_free(workpath);
+				res.err = "Couldn't build FIFO";
+				return res;
+			}
+			int fd = open(workpath, O_RDWR);
+			if (-1 == fd){
+				arcan_mem_free(workpath);
+				res.err = "Couldn't bind FIFO";
+				return res;
+			}
+			res.unlink = workpath;
+			res.fd = fd;
+			return res;
+		}
+		else {
+			res.path = workpath;
+			return res;
+		}
+	}
+
+	int fd = open(workpath, O_RDWR);
+	arcan_mem_free(workpath);
+
+	if (-1 == fd || -1 == fstat(fd, &fi) || S_ISFIFO(fi.st_mode)){
+		close(fd);
+		res.err = "Couldn't open as FIFO";
+		return res;
+	}
+
+	res.fd = fd;
+	return res;
+}
+
+static struct pathfd build_socket_ipc(char* pathin, bool userns, bool srv)
+{
+	struct pathfd res = {.path = NULL, .fd = -1, .err = NULL};
+	int ns = userns ? RESOURCE_NS_USER : RESOURCE_APPL_TEMP;
+
+		struct sockaddr_un addr = {
+			.sun_family = AF_UNIX
+		};
+
+	if (srv){
+		char* workpath = arcan_find_resource(pathin, ns, ARES_FILE, NULL);
+
+		if (workpath){
+			res.err = "EINVAL: Couldn't create socket";
+			arcan_mem_free(workpath);
+			return res;
+		}
+
+		workpath = arcan_expand_resource(pathin, ns);
+		if (!workpath){
+			res.err = "EINVAL: Couldn't build socket file";
+			return res;
+		}
+
+		struct sockaddr_un addr = {
+			.sun_family = AF_UNIX
+		};
+		size_t lim = COUNT_OF(addr.sun_path);
+		if (strlen(workpath) > lim - 1){
+			res.err = "ENAMETOOLONG: expanded socket doesn't fit sockaddr";
+			arcan_mem_free(workpath);
+			return res;
+		}
+		snprintf(addr.sun_path, lim, "%s", workpath);
+
+		res.fd = socket(AF_UNIX, SOCK_STREAM, 0);
+		if (-1 == res.fd){
+			res.err = "EPERM: couldn't allocate socket";
+			arcan_mem_free(workpath);
+			return res;
+		}
+		fchmod(res.fd, S_IRWXU);
+
+		if (-1 == bind(res.fd, (struct sockaddr*) &addr, sizeof(addr))){
+			close(res.fd);
+			arcan_mem_free(workpath);
+			res.fd = -1;
+			res.err = "ESOCKET: couldn't bind socket";
+			return res;
+		}
+
+/* this one takes a different metatable to handle accept on connect */
+		listen(res.fd, 5);
+		res.unlink = workpath;
+		res.metatable = "nonblockIOs";
+		res.wrmode = O_RDWR;
+	}
+	else {
+		char* workpath = arcan_find_resource(pathin, ns, ARES_FILE, NULL);
+
+		if (!workpath){
+			res.err = "EEXIST: Couldn't connect to socket";
+			return res;
+		}
+
+		res.fd = alt_nbio_socket(workpath, ns, &res.unlink);
+		res.wrmode = O_RDWR;
+		res.metatable = "nonblockIO";
+
+		if (-1 == res.fd){
+			res.err = "EPERM: Couldn't bind to socket";
+		}
+
+		arcan_mem_free(workpath);
+	}
+
+	return res;
+}
+
+static struct pathfd build_new_file(char* path, bool userns)
+{
+	struct pathfd res = {
+		.path = NULL,
+		.fd = -1,
+		.err = NULL,
+		.metatable = "nonblockIO",
+		.wrmode = O_RDWR
+	};
+	int ns = userns ? RESOURCE_NS_USER : RESOURCE_APPL_TEMP;
+
+	char* userpath = arcan_find_resource(
+		path, ns, ARES_FILE | ARES_CREATE, &res.fd);
+
+#ifdef WANT_ARCAN_BASE
+	if (lua_debug_level){
+		arcan_warning("find_resource:ns=%d:%s\n", ns, path ? path : "[null]");
+	}
+#endif
+
+	if (!path){
+		res.err = "Couldn't create file in namespace";
+	}
+	else
+		arcan_mem_free(userpath);
+
+	return res;
+}
+
+static struct pathfd open_existing_file(char* path, bool userns)
+{
+	struct pathfd res = {
+		.path = NULL,
+		.err = NULL,
+		.fd = -1,
+		.wrmode = O_RDONLY,
+		.metatable = "nonblockIO"
+	};
+
+	int ns = userns ? RESOURCE_NS_USER : DEFAULT_USERMASK;
+	char* cpath = arcan_find_resource(path, ns, ARES_FILE, &res.fd);
+
+#ifdef WANT_ARCAN_BASE
+	if (lua_debug_level){
+		arcan_warning(
+			"find_resource:ns=%d:%s=%s\n", ns, path, cpath ? cpath : "[null]");
+	}
+#endif
+
+	if (!cpath){
+		res.err = "Couldn't find file";
+	}
+
+	arcan_mem_free(cpath);
+	return res;
+}
+
 int alt_nbio_open(lua_State* L)
 {
 	LUA_TRACE("open_nonblock");
-
-	const char* metatable = "nonblockIO";
-	char* unlink_fn = NULL;
+	struct pathfd pfd;
 
 	int wrmode = luaL_optbnumber(L, 2, 0) ? O_WRONLY : O_RDONLY;
-	bool fifo = false, ignerr = false, use_socket = false;
-	char* path;
-	int fd;
+	bool userns = false;
 
-/* nonblock-io write to/from an explicit vid */
+/* nonblock-io write to/from an explicit vid,
+ * this might also be opening a hash to/from an existing a12 monitor */
 #ifdef WANT_ARCAN_BASE
 	if (lua_type(L, 1) == LUA_TNUMBER){
 		int rv = opennonblock_tgt(L, wrmode == O_WRONLY);
@@ -979,164 +1179,48 @@ int alt_nbio_open(lua_State* L)
 	}
 #endif
 
-	int namespace = RESOURCE_APPL_TEMP;
-	const char* str = luaL_checkstring(L, 1);
-	if (str[0] == '<'){
-		fifo = true;
-		str++;
-	}
-	else if (str[0] == '='){
-		use_socket = true;
-		str++;
-	}
-	else {
-		size_t i = 0;
-		for (;str[i] && isalnum(str[i]); i++);
-		if (str[i] == ':' && str[i+1] == '/'){
-			namespace = RESOURCE_NS_USER;
-		}
+	char* str = strdup(luaL_checkstring(L, 1));
+
+	size_t i = 0;
+	for (;str[i] && isalnum(str[i]); i++);
+	if (str[i] == ':' && str[i+1] == '/'){
+		userns = RESOURCE_NS_USER;
 	}
 
-/* note on file-system races: it is an explicit contract that the namespace
- * provided for RESOURCE_APPL_TEMP is single- user (us) only. Anyhow, this
- * code turned out a lot messier than needed, refactor when time permits. */
-	if (wrmode == O_WRONLY){
-		struct stat fi;
-		path = arcan_find_resource(str, namespace, ARES_FILE, NULL);
+	if (str[0] == '<')
+		pfd = build_fifo_ipc(str+1, userns, wrmode == O_WRONLY);
+	else if (str[0] == '=')
+		pfd = build_socket_ipc(str+1, userns, wrmode != O_WRONLY);
+	else if (wrmode == O_WRONLY)
+		pfd = build_new_file(str, userns);
+	else
+		pfd = open_existing_file(str, userns);
 
-/* we require a zap_resource call if the file already exists, except for in
- * the case of a fifo dst- that we can open in (w) mode */
-		bool dst_fifo = (path && -1 != stat(path, &fi) && S_ISFIFO(fi.st_mode));
-		if (!dst_fifo && (path || !(path = arcan_expand_resource(str, namespace)))){
-			arcan_warning("open_nonblock(), refusing to open "
-				"existing file for writing\n");
-			arcan_mem_free(path);
+	free(str);
 
-			LUA_ETRACE("open_nonblock", "write on already existing file", 0);
-		}
-
-		int fl = O_NONBLOCK | O_WRONLY | O_CLOEXEC;
-		if (fifo){
-/* this is susceptible to the normal race conditions, but we also expect
- * APPL_TEMP to be mapped to a 'safe' path */
-			if (-1 == mkfifo(path, S_IRWXU)){
-				if (errno != EEXIST || -1 == stat(path, &fi) || !S_ISFIFO(fi.st_mode)){
-					arcan_warning("open_nonblock(): mkfifo (%s) failed\n", path);
-					LUA_ETRACE("open_nonblock", "mkfifo failed", 0);
-				}
-			}
-			unlink_fn = strdup(path);
-			ignerr = true;
-		}
-		else
-			fl |= O_CREAT;
-
-/* failure to open fifo can be expected, then opening will be deferred */
-		fd = open(path, fl, S_IRWXU);
-		if (-1 != fd && fifo && (-1 == fstat(fd, &fi) || !S_ISFIFO(fi.st_mode))){
-			close(fd);
-			LUA_ETRACE("open_nonblock", "opened file not fifo", 0);
-		}
-	}
-/* recall, socket binding is supposed to go to a 'safe' namespace, so
- * filesystem races are less than a concern than normally */
-	else if (use_socket){
-		struct sockaddr_un addr = {
-			.sun_family = AF_UNIX
-		};
-		size_t lim = COUNT_OF(addr.sun_path);
-		path = arcan_find_resource(str, namespace, ARES_FILE, NULL);
-		if (path || !(path = arcan_expand_resource(str, namespace))){
-			arcan_warning("open_nonblock(), refusing to overwrite file\n");
-			LUA_ETRACE("open_nonblock", "couldn't create socket", 0);
-		}
-
-		if (strlen(path) > lim - 1){
-			arcan_warning("open_nonblock(), socket path too long\n");
-			LUA_ETRACE("open_nonblock", "socket path too long", 0);
-		}
-		snprintf(addr.sun_path, lim, "%s", path);
-
-		metatable = "nonblockIOs";
-
-		fd = socket(AF_UNIX, SOCK_STREAM, 0);
-		if (-1 == fd){
-			arcan_warning("open_nonblock(): couldn't create socket\n");
-			arcan_mem_free(path);
-			LUA_ETRACE("open_nonblock", "couldn't create socket", 0);
-		}
-		fchmod(fd, S_IRWXU);
-
-		alt_nbio_nonblock_cloexec(fd, true);
-		int rv = bind(fd, (struct sockaddr*) &addr, sizeof(addr));
-		if (-1 == rv){
-			close(fd);
-			arcan_mem_free(path);
-			arcan_warning(
-				"open_nonblock(): bind (%s) failed: %s\n", path, strerror(errno));
-			LUA_ETRACE("open_nonblock", "couldn't bind socket", 0);
-		}
-		listen(fd, 5);
-		unlink_fn = path;
-		path = NULL; /* don't mark as pending */
-	}
-	else {
-retryopen:
-		path = arcan_find_resource(str, namespace, ARES_FILE, NULL);
-
-/* fifo and doesn't exist? create */
-		if (!path){
-			if (fifo && (path = arcan_expand_resource(str, namespace))){
-				if (-1 == mkfifo(path, S_IRWXU)){
-					arcan_warning("open_nonblock(): mkfifo (%s) failed\n", path);
-					LUA_ETRACE("open_nonblock", "mkfifo failed", 0);
-				}
-				goto retryopen;
-			}
-			else{
-				LUA_ETRACE("open_nonblock", "file does not exist", 0);
-			}
-		}
-/* normal file OR socket */
-		else{
-			fd = open(path, O_NONBLOCK | O_CLOEXEC | O_RDONLY);
-
-/* socket, 'connect mode' */
-			if (-1 == fd && errno == ENXIO){
-				fd = alt_nbio_socket(path, namespace, &unlink_fn);
-				wrmode = O_RDWR;
-			}
-		}
-
-		arcan_mem_free(path);
-		path = NULL;
-	}
-
-	if (fd < 0 && !ignerr){
-		arcan_mem_free(path);
-		LUA_ETRACE("open_nonblock", "couldn't open file", 0);
+	if (pfd.err){
+		LUA_ETRACE("open_nonblock", pfd.err, 0);
 	}
 
 	struct nonblock_io* conn = arcan_alloc_mem(
 			sizeof(struct nonblock_io),
 			ARCAN_MEM_BINDING, ARCAN_MEM_BZERO, ARCAN_MEMALIGN_NATURAL);
 
-	conn->fd = fd;
-	conn->canary_pre = 0xfeedface;
-	conn->canary_post = 0xfacefeed;
+	conn->fd = pfd.fd;
+	alt_nbio_nonblock_cloexec(pfd.fd, true);
 
 /* this little crutch was better than differentiating the userdata as the
  * support for polymorphism there is rather clunky */
-	conn->mode = wrmode;
-	conn->pending = path;
-	conn->unlink_fn = unlink_fn;
+	conn->mode = pfd.wrmode;
+	conn->pending = pfd.path;
+	conn->unlink_fn = pfd.unlink;
 	conn->data_handler = LUA_NOREF;
 	conn->write_handler = LUA_NOREF;
 
 	uintptr_t* dp = lua_newuserdata(L, sizeof(uintptr_t));
 	*dp = (uintptr_t) conn;
 
-	luaL_getmetatable(L, metatable);
+	luaL_getmetatable(L, pfd.metatable);
 	lua_setmetatable(L, -2);
 
 	LUA_ETRACE("open_nonblock", NULL, 1);
@@ -1319,13 +1403,12 @@ bool alt_nbio_import(
 	*dp = (uintptr_t) nbio;
 
 	*nbio = (struct nonblock_io){
-		.canary_pre = 0xfeedface,
 		.fd = fd,
 		.mode = mode,
+		.lfch = '\n',
 		.unlink_fn = (unlink_fn ? *unlink_fn : NULL),
 		.write_handler = LUA_NOREF,
 		.data_handler = LUA_NOREF,
-		.canary_post = 0xfacefeed
 	};
 
 	if (out)