+++ src/shmif/tui/lua/tui_lua_glob.h
@@ -1,0 +1,6 @@
+#ifndef HAVE_TUI_LUA_GLOB
+#define HAVE_TUI_LUA_GLOB
+
+int tui_glob(lua_State* L);
+
+#endif